feat: update get product from stock service

+++ stock/internal/entity/stock.go
@@ -4,5 +4,5 @@ import "github.com/google/uuid"
 
 type Stock struct {
 	ProductID uuid.UUID `json:"id"`
-	Quantity  int32     `json:"quantity"`
+	Quantity  int64     `json:"quantity"`
 }
